Use MUnit in QuietTimeoutStageSuite
