+++ core/src/test/scala/org/http4s/blaze/pipeline/stages/TimeoutHelpers.scala
@@ -16,25 +16,24 @@
 
 package org.http4s.blaze.pipeline.stages
 
-import org.specs2.mutable.Specification
-import org.http4s.blaze.pipeline.{Command, LeafBuilder}
 import java.nio.ByteBuffer
 import java.nio.charset.StandardCharsets
 
-import org.specs2.matcher.MatchResult
+import munit.FunSuite
+import org.http4s.blaze.pipeline.{Command, LeafBuilder}
 
 import scala.concurrent.{Await, Future}
 import scala.concurrent.duration._
 
-abstract class TimeoutHelpers extends Specification {
+abstract class TimeoutHelpers extends FunSuite {
   def genDelayStage(timeout: Duration): TimeoutStageBase[ByteBuffer]
 
   def newBuff: ByteBuffer = ByteBuffer.wrap("Foo".getBytes(StandardCharsets.UTF_8))
 
-  def checkBuff(buff: ByteBuffer): MatchResult[Any] =
-    StandardCharsets.UTF_8.decode(buff).toString should_== "Foo"
+  def checkBuff(buff: ByteBuffer): Unit =
+    assertEquals(StandardCharsets.UTF_8.decode(buff).toString, "Foo")
 
-  def checkFuture(f: Future[ByteBuffer], timeout: Duration = 2.seconds): MatchResult[Any] = {
+  def checkFuture(f: Future[ByteBuffer], timeout: Duration = 2.seconds): Unit = {
     val r = Await.result(f, timeout)
     checkBuff(r)
   }
